+++ src/wizards/SumWiz.cpp
@@ -93,7 +93,7 @@ void SumWiz::do_layout() {
   grid_sizer_2->Add(20, 20, 0, 0);
   grid_sizer_2->Add(sizer_2, 1, wxALIGN_LEFT, 0);
   grid_sizer_1->Add(grid_sizer_2, 1, wxEXPAND, 0);
-  grid_sizer_1->Add(static_line_1, 0, wxEXPAND | wxLEFT | wxRIGHT, 2);
+  grid_sizer_1->Add(static_line_1, 0, static_cast<int>(wxEXPAND) | wxLEFT | wxRIGHT, 2);
   sizer_1->Add(button_1, 0, wxALL, 5);
   sizer_1->Add(button_2, 0, wxALL, 5);
   grid_sizer_1->Add(sizer_1, 1, wxALIGN_RIGHT, 0);
